feat: add graph editor toolbar
